Refactor CodeEditorWorkspace and WizardShell components: streamline imports, enhance mobile preview button, and update page metadata for improved user experience.
