+++ app/create/code/page.tsx
@@ -2,7 +2,7 @@ import { CodeEditorWorkspace } from "@/app/components/wizard/CodeEditorWorkspace
 import { Metadata } from "next"
 
 export const metadata: Metadata = {
-  title: "Code Preview — FlutterInit",
+  title: "Code Preview | FlutterInit",
   description: "Browse your generated Flutter scaffold in a VS Code–style editor.",
   alternates: {
     canonical: "https://flutterinit.com/create/code",
